added option to customize GNR sources

+++ data_processing/name_resolver.py
@@ -6,7 +6,7 @@ sys.path.append(os.path.dirname(os.path.dirname(__file__)))
 import glob
 import logging
 from enum import Enum
-from typing import List
+from typing import List, Optional
 import numpy as np
 import pandas as pd
 from taxon_names_resolver import Resolver
@@ -21,10 +21,6 @@ class NameResolutionMethod(str, Enum):
     TNRS = "tnrs"  # https://tnrs.biendata.org/
     GNR = "gnr"  # https://resolver.globalnames.org/
     WFO = "wfo"  # https://github.com/rizqirizqi/scientific-name-scraper
-
-
-name_resolution_result_fields = ["query", "matched_name", "has_mismatches", "score", "taxon_rank", "name_sources"]
-
 
 class NameResolver:
     def __init__(self, method: NameResolutionMethod = "gnr", aux_dir: str = os.getcwd()):
@@ -42,14 +38,14 @@ class NameResolver:
         return input_dir, output_dir, work_dir
 
     @staticmethod
-    def _get_input_batches(items: List[str], write_dir: str, batch_size: int = 200):
+    def _get_input_batches(items: List[str], write_dir: str, batch_size: int = 20):
         input_batches = [items[i: i + batch_size] for i in range(0, len(items), batch_size)]
         for i in range(len(input_batches)):
             input_path = f"{write_dir}/batch_{i}.csv"
-            pd.Series(input_batches[i]).to_csv(input_path, header=["species_name"])
+            pd.Series(input_batches[i]).to_csv(input_path, header=["species_name"], index=False)
         logger.info(f"generated {len(input_batches)} input batches of size {batch_size}")
 
-    def _exec_name_resolution(self, input_dir: str, output_dir: str, work_dir: str):
+    def _exec_name_resolution(self, input_dir: str, output_dir: str, work_dir: str, gnr_data_source: Optional[str] = None):
         job_commands = []
         num_batches = 0
         for path in os.listdir(input_dir):
@@ -61,6 +57,8 @@ class NameResolver:
             if not os.path.exists(f"{batch_output_dir}/output.csv"):
                 func_name = f"exec_{self.method}"
                 command = f"python {os.path.abspath(__file__)} {func_name} {input_path} {output_path}"
+                if self.method == NameResolutionMethod.GNR and gnr_data_source is not None:
+                    command += f" '{gnr_data_source}'"
                 commands = [
                     f"cd {batch_output_dir}",
                     command
@@ -73,37 +71,41 @@ class NameResolver:
     def _unite_batch_data(read_dir: str, output_path: str):
         dfs_paths = glob.glob(read_dir + f"**/*csv", recursive=True)
         dfs = [pd.read_csv(path) for path in dfs_paths]
-        df = pd.concat(dfs)[name_resolution_result_fields]
+        df = pd.concat(dfs)
         df.to_csv(output_path)
 
-    def resolve(self, names: List[str], output_path: str, batch_size: int = 200):
+    def resolve(self, names: List[str], output_path: str, batch_size: int = 20, gnr_data_source: Optional[str] = None):
         input_dir, output_dir, work_dir = self._set_env()
         self._get_input_batches(items=names, write_dir=input_dir, batch_size=batch_size)
-        self._exec_name_resolution(input_dir=input_dir, output_dir=output_dir, work_dir=work_dir)
+        self._exec_name_resolution(input_dir=input_dir, output_dir=output_dir, work_dir=work_dir, gnr_data_source=gnr_data_source)
         self._unite_batch_data(read_dir=output_dir, output_path=output_path)
         shutil.rmtree(input_dir, ignore_errors=True)
         shutil.rmtree(output_dir, ignore_errors=True)
 
     @staticmethod
     def exec_tnrs(input_path: str, output_path: str):
-        names = pd.read_csv(input_path).iloc[:, 1].to_list()
-        ot = OTWebServiceWrapper(api_endpoint='production')
-        response = ot.tnrs_match_names(names=names, do_approximate_matching=True, include_suppressed=True)
-        dfs = []
-        for result in response.response_dict['results']:
-            if len(result['matches']) > 0:
-                sub_df = pd.concat([pd.json_normalize(match) for match in result['matches']])
-                dfs.append(sub_df)
-        df = pd.concat(dfs)
-        df.rename(columns={
-            "search_string": "query",
-            "is_approximate_match": "has_mismatches",
-            "taxon.tax_sources": "name_sources",
-            "taxon.rank": "taxon_rank"
-        }, inplace=True)
-        df.sort_values(["query", "score"], ascending=[True, False], inplace=True)  # select matches with the best score
-        df.drop_duplicates(subset=["query"], keep='first', inplace=True)
-        df.to_csv(output_path)
+        if not os.path.exists(output_path):
+            names = pd.read_csv(input_path).iloc[:, 0].to_list()
+            ot = OTWebServiceWrapper(api_endpoint='production')
+            response = ot.tnrs_match_names(names=names, do_approximate_matching=True, include_suppressed=True)
+            dfs = []
+            for result in response.response_dict['results']:
+                if len(result['matches']) > 0:
+                    sub_df = pd.concat([pd.json_normalize(match) for match in result['matches']])
+                    dfs.append(sub_df)
+            df = pd.concat(dfs)
+            df.rename(columns={
+                "search_string": "query",
+                "is_approximate_match": "has_mismatches",
+                "taxon.tax_sources": "name_sources",
+                "taxon.rank": "taxon_rank",
+
+            }, inplace=True)
+            df["author"] = np.nan
+            df["orig_status"] = df[["taxon.synonyms", "query"]].apply(lambda row: "SYNONYM" if row["query"].capitalize() in row["taxon.synonyms"] else "ACCEPTED", axis=1)
+            df.sort_values(["query", "score"], ascending=[True, False], inplace=True)  # select matches with the best score
+            df.drop_duplicates(subset=["query"], keep='first', inplace=True)
+            df.to_csv(output_path)
 
     @staticmethod
     def exec_wfo(input_path: str, output_path: str):
@@ -114,22 +116,35 @@ class NameResolver:
                      "Fuzzy": "has_mismatches",
                      "scientificName": "matched_name",
                      "taxonRank": "taxon_rank",
-                     "references": "name_sources"})
+                     "references": "name_sources",
+                     "scientificNameAuthorship": "author",
+                     "taxonomicStatus": "orig_status"})
         df["score"] = 1 - (df["Fuzzy.dist"] / 100)
         df.dropna(subset=["query"], inplace=True)
-        df.to_csv()
+        df.to_csv(output_path)
 
     @staticmethod
-    def exec_gnr(input_path: str, output_path: str):
-        resolver = Resolver(input_path)
-        resolver.main()  # to run the search
-        resolver.write()  # to output the csv file
-        df = pd.read_csv(output_path)
+    def exec_gnr(input_path: str, output_path: str, gnr_data_source: Optional[str] = None):
+        base_output_path = f"{os.path.dirname(output_path)}/resolved_names/search_results.csv"
+        if not os.path.exists(base_output_path):
+            orig_dir = os.getcwd()
+            os.chdir(os.path.dirname(output_path))
+            if gnr_data_source:
+                resolver = Resolver(input_path, datasource=[gnr_data_source])
+            else:
+                resolver = Resolver(input_path)
+            resolver.main()  # to run the search
+            resolver.write()  # to output the csv file
+            os.chdir(orig_dir)
+        df = pd.read_csv(base_output_path)
         df.rename(
-            columns={"query_name": "query", "data_source_title": "name_sources", "name_string": "matched_name"},
+            columns={"query_name": "query", "data_source_title": "name_sources", "current_name_string": "matched_name"},
             inplace=True)
-        df["taxon_rank"] = np.nan
+        df["taxon_rank"] = df["classification_path_ranks"].apply(lambda rank: rank.split("|")[-1] if pd.notna(rank) else np.nan)
+        df["orig_status"] = df["matched_name"].apply(lambda name: "SYNONYM" if pd.notna(name) else "ACCEPTED")
+        df["matched_name"].fillna(value=df["canonical_form"].to_dict(), inplace=True)
         df["has_mismatches"] = np.where(df["query"] != df["matched_name"], True, False)
+        df["author"] = np.nan
         df.to_csv(output_path)
